requirements.txt
asgiref==3.8.1
boltons==25.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
filefetcher==0.1.2
GeneLocator==1.1.2
idna==3.10
iniconfig==2.1.0
intervaltree==3.1.0
lmdb==1.6.2
msgpack==1.1.1
numpy==2.3.1
packaging==25.0
pandas==2.3.0
patsy==1.0.1
pluggy==1.6.0
Pygments==2.19.2
pysam==0.23.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.4
scipy==1.16.0
setuptools==78.1.1
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.3
statsmodels==0.14.4
typesense==1.1.1
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
